1.) fix typo in ArabicTextHandler.java 2.) give some extra precision to NumberParser.java 3.) drop some garbage objects in DoubleDocument.java
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@491229 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/parser/NumberParser.java
@@ -45,6 +45,7 @@ public abstract class NumberParser extends AbstractParser {
         switch (current) {
         case '-':
             mantPos = false;
+            // fallthrough
         case '+':
             current = reader.read();
         }
@@ -216,7 +217,7 @@ public abstract class NumberParser extends AbstractParser {
      */
     public static float buildFloat(int mant, int exp) {
         if (exp < -125 || mant == 0) {
-            return 0f;
+            return 0.0f;
         }
 
         if (exp >=  128) {
@@ -233,16 +234,17 @@ public abstract class NumberParser extends AbstractParser {
             mant++;  // round up trailing bits if they will be dropped.
         }
 
-        return (exp > 0) ? mant * pow10[exp] : mant / pow10[-exp];
+        return (float) ((exp > 0) ? mant * pow10[exp] : mant / pow10[-exp]);
     }
 
     /**
-     * Array of powers of ten.
+     * Array of powers of ten. Using double instead of float gives a tiny bit more precision.
      */
-    private static final float[] pow10 = new float [128];
+    private static final double[] pow10 = new double[128];
+
     static {
         for (int i = 0; i < pow10.length; i++) {
-            pow10[i] = (float)Math.pow(10, i);
+            pow10[i] = Math.pow(10, i);
         }
     }
 }
